+++ projects/start-os/web/ui/src/app/routes/portal/routes/services/modals/action-success/types.ts
@@ -3,4 +3,5 @@ import { ActionRes } from 'src/app/services/api/api.types'
 export type ActionResponse = NonNullable<ActionRes>
 type ActionResult = NonNullable<ActionResponse['result']>
 export type SingleResult = ActionResult & { type: 'single' }
+export type MultilineResult = ActionResult & { type: 'multiline' }
 export type GroupResult = ActionResult & { type: 'group' }